+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.7.8...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.7.9...HEAD) - TBD
+
+## [4.7.9](https://github.com/schemathesis/schemathesis/compare/v4.7.8...v4.7.9) - 2025-12-30
 
 ### :rocket: Added